+++ R/aaa-new.R
@@ -94,3 +94,41 @@ metric_direction <- function(x) {
   attr(x, "direction") <- value
   x
 }
+
+#' @noRd
+#' @export
+print.metric <- function(x, ...) {
+  cat(format(x), sep = "\n")
+  invisible(x)
+}
+
+#' @export
+format.metric <- function(x, ...) {
+  first_class <- class(x)[[1]]
+  metric_type <-
+    switch(
+      first_class,
+      "prob_metric" = "probability metric",
+      "class_metric" = "class metric",
+      "numeric_metric" = "numeric metric",
+      "dynamic_survival_metric" = "dynamic survival metric",
+      "static_survival_metric" = "static survival metric",
+      "integrated_survival_metric" = "integrated survival metric",
+      "metric"
+    )
+
+  metric_desc <- "direction: {.field {attr(x, 'direction')}}"
+
+  by_attr <- attr(x, "by")
+  if (!is.null(by_attr)) {
+    metric_desc <-
+      c(
+        metric_desc,
+        ", group-wise on: {.field {as.character(by_attr)}}"
+      )
+  }
+
+  cli::cli_format_method(
+    cli::cli_text(c("A {metric_type} | ", metric_desc))
+  )
+}