+++ NEWS.md
@@ -16,6 +16,8 @@ calculated with `roc_auc_survival()`.
 
 * `demographic_parity()`, `equalized_odds()`, and `equal_opportunity()` are new metrics for measuring model fairness. Each is implemented with the `new_groupwise_metric()` constructor, a general interface for defining group-aware metrics that allows for quickly and flexibly defining fairness metrics with the problem context in mind.
 
+* Added a print method for metrics and metric sets (#435).
+
 * All warnings and errors have been updated to use the cli package for increased clarity and consistency. (#456, #457, #458)
 
 # yardstick 1.2.0